% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{this function outputs a plot of how much a name was given over time.}
\usage{
draw_a_name(the_name, the_sex = c("M", "F"))
}
\arguments{
\item{the_name}{name we want to plot}

\item{the_sex}{specify male or female (optional, both by default)}
}
\value{
a plot
}
\description{
this function outputs a plot of how much a name was given over time.
}
\examples{
draw_a_name("Jean","M")
}
