% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{draws a graph of multiple names over time}
\usage{
draw_names(my_names, the_sex = c("M", "F"))
}
\arguments{
\item{my_names}{list of names}

\item{the_sex}{Male or female, both by default}
}
\value{
a plot
}
\description{
draws a graph of multiple names over time
}
\examples{
draw_names(c("Clément","Bertrand"),"M")

}
