% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{imports all the sheets in an xlsx file as a list of data frames}
\usage{
multi_excel(filepath)
}
\arguments{
\item{filepath}{the path to the file we want to load}
}
\value{
a list of dataframes containing the sheets
}
\description{
imports all the sheets in an xlsx file as a list of data frames
}
\examples{
\dontrun{
multi_excel("/../../some/path")
}
}
