% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{import multiple csv files}
\usage{
my_csv_reader(folder_path)
}
\arguments{
\item{folder_path}{path to a folder from current directory}
}
\value{
path
}
\description{
import multiple csv files
}
\examples{
\dontrun{
my_csv_reader("../../folder/path")
}

}
