% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{saves a dataframe as csv}
\usage{
save_as_csv(df, path, row.names = FALSE, ...)
}
\arguments{
\item{df}{the data frame we want to save}

\item{path}{path to the csv file to be created}

\item{row.names}{row names}

\item{...}{all other parameters for write.csv}
}
\value{
nothing
}
\description{
saves a dataframe as csv
}
\examples{
\dontrun{
my_df <- as.data.frame(c(1,24,3))
save_as_csv(my_df, "..\\\\..\\\\somefolder\\\\save_as_csv_test2.csv")
}

}
